+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Fixed some HK translations
 - Fixed non-UTF filename attachment open and download
 - Fixed save to sent folder behavior - [#98](https://github.com/hkdb/aerion/issues/98)
+- Improved composer formating
 
 
 **v0.1.38 - 03-22-2026**